convert stringKey read exception to `None`
for whatever reason when we read other formats like enumerations we will convert exceptinos to `None`. but for our case classes we have neglected to do so. I believe it's because we assume that if we have the right types then `Apply` method should work, but where we have case classes that have additional validation in the constructor (e.g. require) then deserializing the string key will throw an exception rather than return `None`. among other things, this causes our naptime request handler to throw a 5XX instead of a 4XX when parsing bad values.